# doc/api.yaml
openapi: 3.0.0
paths:
...
  /track:
    post:
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        description: A name, artistId, albumId and duration for the new track
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "201":
          description: The track has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Request body does not contain required fields
      tags: &a3
        - track
    get:
      operationId: TracksController_findAll
      parameters: []
      responses:
        "200":
          description: A tracks has been successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
      tags: *a3
  "/track/{id}":
    get:
      operationId: TracksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a track that exists in the database
          schema:
            type: string
      responses:
        "200":
          description: A track has been successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: A track with given id is invalid (not uuid).
        "404":
          description: A track with given id does not exist.
      tags: *a3
    put:
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a track that exists in the database
          schema:
            type: string
      requestBody:
        required: true
        description: A new name, artistId, albumId and duration for the track
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: A track has been successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: A track with given id is invalid (not uuid).
        "404":
          description: A track with given id does not exist.
      tags: *a3
    delete:
      operationId: TracksController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a track that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A track has been successfully deleted
        "400":
          description: A track with given id is invalid (not uuid).
        "404":
          description: A track with given id does not exist.
      tags: *a3
...
  "/favs/track/{id}":
    post:
      operationId: TrackToFavoritesController_createTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a track that exists in the database
          schema:
            type: string
      responses:
        "201":
          description: A track has been successfully added
        "400":
          description: A track with given id is invalid (not uuid).
        "422":
          description: A track with given id does not exist
      tags: &a6
        - favs/track
    delete:
      operationId: TrackToFavoritesController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a track that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A track has been successfully deleted
        "400":
          description: A track with given id is invalid (not uuid).
        "404":
          description: A track with given id does not exist.
      tags: *a6
...
info:
  title: Home Library Service
  description: Home music library service OpenAPI 3.0
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:4000
  - url: http://localhost
components:
  schemas:
...
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - name
        - duration
    Track:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - name
        - duration
...